backend/requirements.txt
# Mock Trial AI Backend Dependencies
# Per README.md: Python FastAPI with LangGraph

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Data Validation
pydantic>=2.5.0

# AI/ML
openai>=1.12.0
langgraph>=0.0.26
anthropic>=0.40.0
google-genai>=1.0.0

# Vector Database
pinecone>=5.0.0

# Database (Supabase)
supabase>=2.0.0
# Optional: Direct PostgreSQL (if pooler connection works)
# psycopg2-binary>=2.9.9
# sqlalchemy>=2.0.25

# WebSocket support (included in fastapi)
websockets>=12.0

# Audio processing
# wave is part of standard library

# PDF parsing
PyPDF2>=3.0.0

# HTTP client (for fetching case PDFs)
httpx>=0.27.0

# Development
python-dotenv>=1.0.0
